状態などを子コンポーネントに移動
