+++ src/components/ingredients-form.tsx
@@ -1,48 +1,217 @@
-import { IngredientGroup, Unit } from "@/types/index";
+import { Ingredient, IngredientGroup, Unit } from "@/types/index";
+import { useEffect, useMemo, useState } from "react";
 
 type Props = {
-  inputName: string;
-  inputPrice: number | null;
-  setInputName: React.Dispatch<React.SetStateAction<string>>;
-  selectedGroupId: number | null;
-  selectedUnitId: number | null;
-  selectedUnit: Unit | undefined;
-  handleSelectGroupId: (groupId: number | null) => void;
-  handleSelectUnitId: (unitId: number | null) => void;
-  handleChangePrice: (price: number | null) => void;
-  handleSubmitSave: () => void;
-  handleSubmitEdit: () => void;
+  // inputName: string;
+  // inputPrice: number | null;
+  // setInputName: React.Dispatch<React.SetStateAction<string>>;
+  // selectedGroupId: number | null;
+  // selectedUnitId: number | null;
+  // selectedUnit: Unit | undefined;
+  // handleSelectGroupId: (groupId: number | null) => void;
+  // handleSelectUnitId: (unitId: number | null) => void;
+  // handleChangePrice: (price: number | null) => void;
+  // handleSubmitSave: () => void;
+  // handleSubmitEdit: () => void;
+  ingredients: Ingredient[];
+  setIngredients: React.Dispatch<React.SetStateAction<Ingredient[]>>;
+  editingIngredientId: number | null;
+  setEditIngredientId: React.Dispatch<React.SetStateAction<number | null>>;
   ingredientGroups: IngredientGroup[];
   units: Unit[];
-  isEditMode: boolean;
-  isDisabled: boolean;
+  // isEditMode: boolean;
+  // isDisabled: boolean;
 };
-
 
 const IngredientsForm = (props: Props) => {
   //Propsを代入
   const {
-    inputName,
-    inputPrice,
-    setInputName,
-    selectedGroupId,
-    selectedUnitId,
-    selectedUnit,
-    handleSelectGroupId,
-    handleSelectUnitId,
-    handleChangePrice,
-    handleSubmitSave,
-    handleSubmitEdit,
+    // inputName,
+    // inputPrice,
+    // setInputName,
+    // selectedGroupId,
+    // selectedUnitId,
+    // selectedUnit,
+    // handleSelectGroupId,
+    // handleSelectUnitId,
+    // handleChangePrice,
+    // handleSubmitSave,
+    // handleSubmitEdit,
+    ingredients,
+    setIngredients,
+    editingIngredientId = null,
+    setEditIngredientId,
     ingredientGroups,
     units,
-    isEditMode,
-    isDisabled,
+    // isEditMode,
+    // isDisabled,
   } = props;
+
+  //[input]材料名
+  const [inputName, setInputName] = useState<string>("");
+
+  //[select]単位
+  const [selectedUnitId, setSelectedUnitId] = useState<number | null>(null);
+
+  //[select]分類
+  const [selectedGroupId, setSelectedGroupId] = useState<number | null>(null);
+
+  //[input]相場
+  const [inputPrice, setInputPrice] = useState<number | null>(null);
+
+  //[select]単位の状態を変更
+  const handleSelectUnitId = (unitId: number | null) => {
+    setSelectedUnitId(unitId);
+  };
+
+  //[select]分類の状態を変更
+  const handleSelectGroupId = (groupId: number | null) => {
+    setSelectedGroupId(groupId);
+  };
+
+  //[input]相場の状態を変更
+  const handleChangePrice = (price: number | null) => {
+    setInputPrice(price);
+  };
+
+  // フォームを送れる状態かどうか
+  const isDisabled: boolean = useMemo(() => {
+    return (
+      // [input]材料名が入っているか
+      !inputName.trim() ||
+      // [select]単位が選ばれているか
+      selectedUnitId === null ||
+      // [select]分類が選ばれているか
+      selectedGroupId === null
+    );
+  }, [inputName, selectedUnitId, selectedGroupId]);
+
+  //セレクトされている単位情報を取得
+  const selectedUnit: Unit | undefined = useMemo(() => {
+    return units.find((unit) => unit.id === selectedUnitId);
+  }, [units, selectedUnitId]);
 
   //単位の文言を変更する際に使用
   const unitLabel = selectedUnit
     ? `${selectedUnit.amountPerUnit}${selectedUnit.name}`
     : "1単位";
+
+  //フォームの送信（保存）
+  const handleSubmitSave = async () => {
+    //編集モードでないこと
+    if (
+      selectedUnitId === null ||
+      selectedGroupId === null ||
+      editingIngredientId !== null
+    )
+      return;
+
+    // 新しい材料を作成
+    const newIngredient = {
+      // id: nextId.current,
+      name: inputName,
+      ingredientGroupId: selectedGroupId,
+      unitId: selectedUnitId,
+      pricePerUnit: inputPrice,
+    };
+
+    try {
+      const res = await fetch("/api/ingredients", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(newIngredient),
+      });
+
+      const result = await res.json();
+
+      if (!res.ok) {
+        throw new Error(result.error || "登録に失敗しました。");
+      }
+
+      const getRes = await fetch("api/ingredients");
+      const data = await getRes.json();
+      setIngredients(data);
+
+      //フォームの内容をリセット
+      setInputName("");
+      setSelectedUnitId(null);
+      setSelectedGroupId(null);
+      setInputPrice(null);
+    } catch (error) {
+      alert("登録に失敗しました");
+      console.error(error);
+    }
+  };
+
+  //フォームの送信（編集）
+  const handleSubmitEdit = async () => {
+    //編集モードであること
+    if (
+      selectedUnitId === null ||
+      selectedGroupId === null ||
+      editingIngredientId === null
+    )
+      return;
+
+    // 新しい材料を作成
+    const newIngredient = {
+      // id: editingIngredientId,
+      name: inputName,
+      ingredientGroupId: selectedGroupId,
+      unitId: selectedUnitId,
+      pricePerUnit: inputPrice,
+    };
+
+    try {
+      const res = await fetch(`/api/ingredients/${editingIngredientId}`, {
+        method: "PUT",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(newIngredient),
+      });
+
+      const result = await res.json();
+
+      if (!res.ok) {
+        throw new Error(result.error || "更新に失敗しました");
+      }
+
+      const getRes = await fetch("api/ingredients");
+      const data = await getRes.json();
+      setIngredients(data);
+
+      //フォームの内容をリセット
+      setInputName("");
+      setSelectedUnitId(null);
+      setSelectedGroupId(null);
+      setInputPrice(null);
+      setEditIngredientId(null);
+    } catch (error) {
+      alert("更新に失敗しました");
+      console.error(error);
+    }
+  };
+
+  //編集時にフォームに編集対象の情報を初期値としてセット
+  useEffect(() => {
+    if (editingIngredientId !== null) {
+      const editTarget = ingredients.find((i) => i.id === editingIngredientId);
+      if (editTarget) {
+        setInputName(editTarget.name);
+        setSelectedGroupId(editTarget.ingredientGroupId);
+        setSelectedUnitId(editTarget.unitId);
+        setInputPrice(editTarget.pricePerUnit ?? null);
+      }
+    } else {
+      setInputName("");
+      setSelectedGroupId(null);
+      setSelectedUnitId(null);
+      setInputPrice(null);
+    }
+  }, [editingIngredientId, ingredients]);
 
   return (
     // フォーム
@@ -50,7 +219,7 @@ const IngredientsForm = (props: Props) => {
       onSubmit={(e) => {
         e.preventDefault();
         //編集中かどうかで処理をかえる
-        if (isEditMode) {
+        if (editingIngredientId !== null) {
           handleSubmitEdit();
         } else {
           handleSubmitSave();
@@ -58,7 +227,7 @@ const IngredientsForm = (props: Props) => {
       }}
       className="mb-5"
     >
-      <h2>{`材料を${isEditMode ? "編集" : "登録"}`}</h2>
+      <h2>{`材料を${editingIngredientId !== null ? "編集" : "登録"}`}</h2>
 
       {/* [input]材料名 */}
       <label className="flex">
@@ -140,7 +309,7 @@ const IngredientsForm = (props: Props) => {
 
       {/* 送信ボタン */}
       <div className="">
-        {isEditMode ? (
+        {editingIngredientId !== null ? (
           <button
             type="submit"
             disabled={isDisabled}
